+++ PlotData.h
@@ -16,16 +16,16 @@
 class PlotData
 {
 public:
-	PlotData(const ulong len, const int style = FL_SOLID, const int width = 1, const Fl_Color col = FL_BLACK);
+	PlotData(const size_t len, const int style = FL_SOLID, const int width = 1, const Fl_Color col = FL_BLACK);
 	PlotData(const std::vector<double> &x, const std::vector<double> &y, const int style = FL_SOLID, const int width = 1, const Fl_Color col = FL_BLACK);
 	~PlotData();
 
 	void getXlim(double &xmin, double &xmax);
 	void getYlim(double &ymin, double &ymax);
-	void setVal(const ulong idx, const double nx, const double ny);
-	void getVal(const ulong idx, double &nx, double &ny);
+	void setVal(const size_t idx, const double nx, const double ny);
+	void getVal(const size_t idx, double &nx, double &ny);
 
-	const ulong size() { return this->len; }
+	const size_t size() { return this->len; }
 
 public:
 	int style;
@@ -36,7 +36,7 @@ private:
 	std::vector<double> x, y;
 	double xmin, ymin;
 	double xmax, ymax;
-	ulong len;
+	size_t len;
 };
 
 #endif /* PLOTDATA_H_ */
